Use `minidump-synth` for the test
